Chore(package): update to version 0.2.0
add Lib.fillPlaceholders() (same on Bundle) method as a replacement to Bundle.insertLibVersionsRadiosHere() method, which required to be called in a smal <script> at the position where the Radio Inputs were to be inserted.
With fillPlaceholders(), we define a placeholder in the HTML (typically a `<div data-manage-lib="libName" data-manage-versions="versionName1, versionName2"></div>`), then we call the fillPlaceholders() method that will automatically look in the document for these placeholders and fill them with the requested Versions Radio Inputs.
This is easier to read (especially in the HTML), we no longer need to load the library before the HTML, and it is more in line with other libraries (typically jQuery, Bootstrap, etc.).